requirement.txt
setuptools
wheel
ipykernel
numpy
pandas
requests
pyspark
matplotlib
seaborn
findspark
aiohttp
pyarrow
google-cloud-storage
google-cloud-bigquery
pymysql
psycopg2-binary
python-dotenv